requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cache2==0.2.2
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.7
passlib==1.7.4
pendulum==3.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
redis==4.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
